update runners and visualizers

+++ c/include/misc.h
@@ -3,6 +3,6 @@
 
 #include <stddef.h>
 
-unsigned long long allocation(int* tasks, size_t n, int* workers, size_t k);
+float allocation(int* tasks, size_t n, float* workers, size_t k);
 
 #endif  // _MISC_H_
